# platform/neuroforge/openapi.yaml
openapi: 3.1.0
info:
  title: NeuroForge API
  version: 0.8.2
  description: REST API for NeuroForge associative learning, explainable vector recall,
    provenance, document/text ingestion, SearXNG-backed autonomous research, source-grounded
    goal cycles, responsive knowledge graph, model routing, cost controls, Prometheus
    observability, and durable storage/cluster operation.
servers:
- url: http://localhost:8080
components:
  securitySchemes:
    AppKey:
      type: http
      scheme: bearer
    WorkerKey:
      type: http
      scheme: bearer
    ControlReadKey:
      type: http
      scheme: bearer
      description: Scoped read token used by the Control Center for graph/orchestrator status.
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
    MetricsToken:
      type: http
      scheme: bearer
      description: Dedicated Prometheus scrape token. The admin token is also accepted
        as a bearer credential.
    ClusterToken:
      type: apiKey
      in: header
      name: X-Cluster-Token
...
paths:
...
  /api/v1/stats:
    get:
      security:
      - AppKey: []
      - AdminToken: []
      responses:
        '200':
          description: Memory
          HNSW: null
          shard: null
          cluster: null
          maintenance and cost counters: null
...
  /api/v1/integrations/graph/status:
    get:
      security:
      - ControlReadKey: []
      - AppKey: []
      - AdminToken: []
      description: Structural knowledge-graph health and graph/backfill configuration.
      responses:
        '200':
          description: Graph statistics including linked/isolated/multi-linked nodes and connected components.
  /api/v1/integrations/orchestrator/status:
    get:
      security:
      - ControlReadKey: []
      - AppKey: []
      - AdminToken: []
      description: Read-only master/subagent scheduler status, workers, queues and offload settings.
      responses:
        '200': {description: Orchestrator status}
...